depthpro_proj/requirements.txt
numpy>=1.24
Pillow>=10.0
open3d>=0.18
